trika-ai/backend/requirements.txt
# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.3
pydantic-settings>=2.1.0

# LangChain & AI
langchain>=0.1.4
langchain-openai>=0.0.5
langchain-community>=0.0.16
langchain-anthropic>=0.1.1
langgraph>=0.0.20
openai>=1.10.0

# Vector Store
# Using Chroma by default; Pinecone packages removed to avoid Python version conflicts
# chromadb==0.4.22
# pinecone-client==3.1.0
# langchain-pinecone>=0.0.3

# Database
sqlalchemy>=2.0.25

# Security & Rate Limiting
slowapi>=0.1.9

# Production Server
gunicorn>=21.2.0

# Async & Utils
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
duckduckgo-search>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
